Proyectos/Proyecto2/API/requirements.txt
Flask==3.0.0
Flask_Cors==4.0.0
mysql_connector_repackaged==0.3.1
python-dotenv==1.0.0
redis==5.0.1
